Move result into util class
